+++ config/api-config.php
@@ -3,9 +3,13 @@
 return [
     'apiConfig' => [
         'base_uri' => 'http://10.0.2.49/portal/api/',
-        //'base_uri' => 'http://10.0.2.140/',
+        //'base_uri' => 'http://192.168.30.66/',
         //'base_uri' => 'http://103.6.179.214:37210/portal/api/',
         'username' => 'pppippi',
         'password' => 'pppippi',
+
+        //'base_uri' => 'http://59.148.233.70:37210/',
+        //'username' => 'pRimEaPp2017',
+        //'password' => 'b3v8s7um8z4QfVTL',
     ],
 ];