+++ httpdocs/assets/js/module.js
@@ -257,11 +257,11 @@ var BreadCrumb = (function () {
     };
 })();
 
-$(function () {
+//$(function () {
     Vue.config.devtools = true;
     WebFont.load({
         google: {
             families: ['Open Sans:300,300i,400,400i,600,600i,700,700i,800,800i', 'Material Icons']
         }
     });
-});
+//});